feat: enhance image analysis service with image retrieval and probability extraction

+++ server/data/imageanalysis/src/main/kotlin/com/portafolio/vientosdelsur/data/imageanalysis/mapper/PredictionMapper.kt
@@ -6,7 +6,7 @@ import com.portafolio.vientosdelsur.domain.imageanalysis.ResultTag
 
 fun PredictionDto.toImageAnalysisResult() = ImageAnalysisResult(
     tagName = fromTagName(tagName),
-    probability = probability.toFloat()
+    probability = probability
 )
 
 fun fromTagName(tagName: String) =
