+++ server/domain/imageanalysis/src/main/kotlin/com/portafolio/vientosdelsur/domain/imageanalysis/ImageAnalysisResult.kt
@@ -2,5 +2,5 @@ package com.portafolio.vientosdelsur.domain.imageanalysis
 
 data class ImageAnalysisResult(
     val tagName: ResultTag,
-    val probability: Float
+    val probability: Double
 )
